2nd exercise done, refactored function/variable names
