+++ app/middlewares/cache-control.js
@@ -1,0 +1,6 @@
+function cacheControl(req, res, next) {
+  res.set("Cache-Control", "no-cache");
+  next();
+}
+
+module.exports = cacheControl;
