+++ app/middlewares/check-mime-type.js
