+++ app/middlewares/has-file.js
