+++ app/middlewares/media-error-handler.js
