+++ app/middlewares/middlewares.js
@@ -1,5 +1,13 @@
-const checkMimeType = require("./checkmitype");
-const isUpload = require("./isUpload");
-const mediaError = require("./mediaerror");
+const checkMimeType = require("./check-mime-type");
+const isUpload = require("./has-file");
+const mediaError = require("./media-error-handler");
+const cacheControl = require("./cache-control");
+const getCurrentTime = require("./get-current-time");
 
-module.exports = { checkMimeType, isUpload, mediaError };
+module.exports = {
+  checkMimeType,
+  isUpload,
+  mediaError,
+  cacheControl,
+  getCurrentTime,
+};
